Interface and generation

+++ managers/timeslotsmanager.cpp
@@ -1,15 +1,16 @@
 #include "timeslotsmanager.h"
 #include "ui_timeslotsmanager.h"
 
-TimeslotsManager::TimeslotsManager(QSqlDatabase *db, QWidget *parent) :
+TimeslotsManager::TimeslotsManager(QSqlDatabase *db, QDate date, QWidget *parent) :
     QDialog(parent),
     ui(new Ui::TimeslotsManager)
 {
     //UI
     ui->setupUi(this);
 
-    //DB
+    //DB and date
     m_db = db;
+    m_date = date;
 
     //Initialise days
     days << "" << "Lundi" << "Mardi" << "Mercredi" << "Jeudi" << "Vendredi" << "Samedi";
@@ -19,6 +20,8 @@ TimeslotsManager::TimeslotsManager(QSqlDatabase *db, QWidget *parent) :
     connect(ui->listKholleurs, SIGNAL(itemSelectionChanged()), this, SLOT(onSelection_change()));
     connect(ui->addButton, SIGNAL(clicked()), this, SLOT(addTimeslot()));
     connect(ui->deleteButton, SIGNAL(clicked()), this, SLOT(deleteTimeslot()));
+    connect(ui->copyButton, SIGNAL(clicked()), this, SLOT(copyTimeslots()));
+    connect(ui->copyAllButton, SIGNAL(clicked()), this, SLOT(copyAllTimeslots()));
 }
 
 TimeslotsManager::~TimeslotsManager()
@@ -66,8 +69,12 @@ void TimeslotsManager::update_list_timeslots(int id_kholleur) {
 
     //Prepare query
     QSqlQuery query(*m_db);
-    query.prepare("SELECT id, time_start, time, time_end, id_kholleurs, id_day FROM tau_timeslots WHERE id_kholleurs=:id_kholleurs ORDER BY id_day, time");
+    query.prepare("SELECT id, time_start, time, time_end, id_kholleurs, date, pupils FROM tau_timeslots "
+                  "WHERE id_kholleurs=:id_kholleurs AND date>=:monday_date AND date<=:sunday_date "
+                  "ORDER BY date, time");
     query.bindValue(":id_kholleurs", id_kholleur);
+    query.bindValue(":monday_date", m_date.toString("yyyy-MM-dd"));
+    query.bindValue(":sunday_date", m_date.addDays(6).toString("yyyy-MM-dd"));
     query.exec();
 
     //Treat
@@ -78,10 +85,13 @@ void TimeslotsManager::update_list_timeslots(int id_kholleur) {
         ts->setTime(QTime::fromString(query.value(2).toString(), "h:mm:ss"));
         ts->setTime_end(QTime::fromString(query.value(3).toString(), "h:mm:ss"));
         ts->setId_kholleurs(query.value(4).toInt());
-        ts->setId_day(query.value(5).toInt());
+        ts->setDate(QDate::fromString(query.value(5).toString(), "yyyy-M-d"));
+        ts->setPupils(query.value(6).toInt());
         queue_displayedTimeslots.enqueue(ts);
 
-        QListWidgetItem *item = new QListWidgetItem(days[ts->getId_day()] + " : " + ts->getTime_start().toString("HH:mm") + " >> " + ts->getTime().toString("HH:mm") + " >> " + ts->getTime_end().toString("HH:mm"), ui->listTimeslots);
+        QListWidgetItem *item = new QListWidgetItem(days[ts->getDate().dayOfWeek()] + " : "
+                + ts->getTime_start().toString("HH:mm") + " >> " + ts->getTime().toString("HH:mm") + " >> " + ts->getTime_end().toString("HH:mm")
+                + ", " + QString::number(ts->getPupils()) + (ts->getPupils() <= 1 ? " élève" : " élèves"), ui->listTimeslots);
         item->setData(Qt::UserRole, (qulonglong) ts);
     }
 }
@@ -93,6 +103,7 @@ void TimeslotsManager::free_timeslots() {
 }
 
 void TimeslotsManager::onSelection_change() {
+    ui->listTimeslots->clear();
     QList<QListWidgetItem*> selection = ui->listKholleurs->selectedItems();
 
     if(selection.length() <= 0) {
@@ -104,6 +115,7 @@ void TimeslotsManager::onSelection_change() {
         ui->listTimeslots->setEnabled(false);
         ui->addButton->setEnabled(false);
         ui->deleteButton->setEnabled(false);
+        ui->copyButton->setEnabled(false);
         return;
     }
 
@@ -115,6 +127,7 @@ void TimeslotsManager::onSelection_change() {
     ui->listTimeslots->setEnabled(true);
     ui->addButton->setEnabled(true);
     ui->deleteButton->setEnabled(true);
+    ui->copyButton->setEnabled(true);
 
     update_list_timeslots(((Kholleur*)selection[0]->data(Qt::UserRole).toULongLong())->getId());
 }
@@ -130,8 +143,9 @@ void TimeslotsManager::addTimeslot() {
 
     //Create timeslot
     Timeslot* ts = new Timeslot();
-    ts->setId_day(ui->comboBox->currentIndex() + 1);
+    ts->setDate(m_date.addDays(ui->comboBox->currentIndex()));
     ts->setId_kholleurs(k->getId());
+    ts->setPupils(k->getPupils());
 
     QTime time = ui->timeEdit->time();
     ts->setTime(time);
@@ -140,12 +154,13 @@ void TimeslotsManager::addTimeslot() {
 
     //Add to DB
     QSqlQuery query(*m_db);
-    query.prepare("INSERT INTO tau_timeslots(time_start, time, time_end, id_kholleurs, id_day) VALUES(:time_start, :time, :time_end, :id_kholleurs, :id_day)");
+    query.prepare("INSERT INTO tau_timeslots(time_start, time, time_end, id_kholleurs, date, pupils) VALUES(:time_start, :time, :time_end, :id_kholleurs, :date, :pupils)");
     query.bindValue(":time_start", ts->getTime_start().toString("HH:mm:ss"));
     query.bindValue(":time", ts->getTime().toString("HH:mm:ss"));
     query.bindValue(":time_end", ts->getTime_end().toString("HH:mm:ss"));
     query.bindValue(":id_kholleurs", ts->getId_kholleurs());
-    query.bindValue(":id_day", ts->getId_day());
+    query.bindValue(":date", ts->getDate().toString("yyyy-MM-dd"));
+    query.bindValue(":pupils", ts->getPupils());
     query.exec();
 
     delete ts;
@@ -162,18 +177,62 @@ void TimeslotsManager::deleteTimeslot() {
 
     Timeslot* ts = (Timeslot*) selection[0]->data(Qt::UserRole).toULongLong();
 
-    //No confirmation needed, I think
-
-    //Query
-    QSqlQuery query(*m_db);
-    query.prepare("DELETE FROM tau_timeslots WHERE id=:id");
-    query.bindValue(":id", ts->getId());
-    query.exec();
-
-    //Update
     QList<QListWidgetItem*> kholleurs = ui->listKholleurs->selectedItems();
-    if(kholleurs.length() > 0) {
-        update_list_timeslots(((Kholleur*)kholleurs[0]->data(Qt::UserRole).toULongLong())->getId());
+    if(kholleurs.length() <= 0) {
+        QMessageBox::critical(this, "Erreur", "Aucun kholleur n'est sélectionné. Comment est-ce possible ?");
+        return;
     }
 
+    //Confirmation
+    int res = QMessageBox::warning(this, "Suppression en cours", "Vous êtes sur le point de supprimer un horaire de kholle "
+                                                                 "du kholleur <strong>" + ((Kholleur*)kholleurs[0]->data(Qt::UserRole).toULongLong())->getName() + "</strong> "
+                                                                 "ainsi que les <strong>kholles</strong> associées. <br /> Voulez-vous continuer ?", QMessageBox::Yes | QMessageBox::No);
+
+    if(res == QMessageBox::Yes) {
+        //Query
+        QSqlQuery query(*m_db);
+        query.prepare("DELETE FROM tau_kholles WHERE id_timeslots=:id_timeslots");
+        query.bindValue(":id_timeslots", ts->getId());
+        query.exec();
+        query.prepare("DELETE FROM tau_timeslots WHERE id=:id");
+        query.bindValue(":id", ts->getId());
+        query.exec();
+
+        //Update
+        update_list_timeslots(((Kholleur*)kholleurs[0]->data(Qt::UserRole).toULongLong())->getId());
+
+    }
+
+}
+
+void TimeslotsManager::copyTimeslots() {
+    //Get selection
+    QList<QListWidgetItem*> selection = ui->listKholleurs->selectedItems();
+
+    if(selection.length() <= 0) {
+        QMessageBox::critical(this, "Erreur", "Veuillez sélectionner un kholleur.");
+        return;
+    }
+    Kholleur* k = (Kholleur*) selection[0]->data(Qt::UserRole).toULongLong();
+
+    //Open dialog
+    CopyTimeslots dialog(m_db, m_date, false, k->getId());
+    dialog.exec();
+
+    //Update
+    update_list_timeslots(k->getId());
+}
+
+void TimeslotsManager::copyAllTimeslots() {
+    //Open dialog with all = true
+    CopyTimeslots dialog(m_db, m_date);
+    dialog.exec();
+
+    //Get selection
+    QList<QListWidgetItem*> selection = ui->listKholleurs->selectedItems();
+
+    //Update if necessary
+    if(selection.length() > 0) {
+        update_list_timeslots(((Kholleur*) selection[0]->data(Qt::UserRole).toULongLong())->getId());
+    }
 }
